requirements.txt
boto3==1.35.92
botocore==1.35.92
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
dnspython==2.7.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
numpy==2.0.2
pafy==0.5.5
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
scrapetube==2.5.1
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20